+++ src/locales/en/footer.js
@@ -15,4 +15,5 @@ export default {
   footerTwitterAlt: 'Twitter Logo',
   footerOcticons: 'Octicons&nbsp;&copy;&nbsp;2021&nbsp;GitHub,&nbsp;Inc.',
   footerCopyright: '&copy;&nbsp;2021&nbsp;Visiosto&nbsp;oy',
+  footerMadeBy: 'The site is rendered through the lens of <a>Visiosto</a>',
 };